+++ app/database/migrations/2013_09_03_084147_create_menus_tables.php
@@ -45,7 +45,7 @@ class CreateMenusTables extends Migration
             $table->increments('id')->unsigned();
             $table->integer('menu_id')->unsigned();
             $table->integer('page_id')->unsigned()->nullable();
-            $table->integer('parent')->unsigned()->default(0);
+            $table->integer('menulink_id')->unsigned()->default(0);
             $table->integer('position')->unsigned()->default(0);
             $table->string('target', 10)->nullable();
             $table->string('module_name', 50)->nullable();